backend/pyproject.toml
[project]
name = "policyops-api"
version = "0.1.0"
description = "PolicyOps Copilot API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "python-json-logger>=2.0.7",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "numpy>=1.26.4",
  "scikit-learn>=1.5.0",
  "httpx>=0.27.0",
  "langgraph>=0.2.40",
]


[project.optional-dependencies]
test = ["pytest>=8.2.0"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
